t/BioFetch.t: don't use eval to decide whether to skip tests.
eval was being used and tests skipped if it failed.  The idea was
probably to avoid false positives if there was any problem on the web
connection.  However, this also meant that any error, even errors in
our code, skipped the tests.  This made the tests quite pointless.

We already test for internet connection at the start of the test so
lets rely on that.